Commandes de création opérationnelles
Modify Application::Run() in order to allow te user to create : Segment,
Rectangle, Polygone, UnionObject and InterObject.
BUT OI & OR will work as soon as the constructor of InterObject and
UnionObject will be implemented. For the moment, there is just an error
poping when we try to use OI and OR.