LOAD opérationnelle, UTF8, meilleurs gestion mémoire
Add class LoadCompositeCommand in order to correctly load Object from a
file, and undo it.
Add a method Clone( ) to every commands in order to manage memory more
efficiently (no more duplicated and unused pointers to commands). Btw
the gestion of "who delete what ?" is now clearer.
Modify slightly Segment::print : it now no longer prints a \n.
Modify FileManager::Load in order to correctly load Objects from a file.
Modify Application to allow the user to use the command LOAD, and
prevent memory leaks.

A simple file "test.txt" is here to show that the load command seems to
work well (more tests needed). It describes the easiest way to write
down a figure in text format.

The SAVE command will be implemented as soon as Object::Print() will be
ready.